+++ R/contrast.R
@@ -384,7 +384,7 @@ contrast.emmGrid = function(object, method = "eff", interaction = FALSE,
         else {
             misc$initMesg = c(misc$initMesg, 
                               paste("Note: contrasts are still on the", misc$orig.tran, "scale"))
-            message("Use 'contrast(regrid(object), ...)' to obtain contrasts on the response scale")
+            message("Note: Use 'contrast(regrid(object), ...)' to obtain contrasts of back-transformed estimates")
             misc$tran = misc$tran.mult = misc$tran.offset = NULL
         }
     }