+++ NEWS.md
@@ -2,14 +2,23 @@
 
 emmeans 1.4.1.009xxx
 -------------
-  * Bug fixes: `regrid` ignored offsets with Bayesian models; `emtrends()` did
-    not supply `options` and `misc` arguments to `emm_basis()` (#143)
+
+  * First argument of `emtrends()` is now `object`, not `model`, to avoid
+    potential mis-matching of the latter with optional `mode` argument
+  * `emtrends()` now uses more robust and efficient code whereby a single
+    reference grid is constructed containing all needed values of `var`. The old
+    version could fail, e.g., in cases where the reference grid involves
+    post-processing. (#145)
   * Added `scale` argument to `contrast()`
   * Added new `"identity"` contrast method
   * New `eff_size()` function for Cohen effect sizes
+  * Bug fixes: `regrid` ignored offsets with Bayesian models; `emtrends()` did
+    not supply `options` and `misc` arguments to `emm_basis()` (#143)
+
 
 emmeans 1.4.1
 -------------
+
   * Added non-estimability infrastructure for Bayesian models, `stanreg`
     in particular (#114)
   * Added `max.degree` argument in `emtrends()` making it possible to